Divide List component source into descrete file
List に関係する部分を Apollo 周り含めて List.jsx に切り出してみた
