+++ src/List.jsx
@@ -1,9 +1,60 @@
 import React from 'react';
+import {
+  ApolloClient,
+  createNetworkInterface,
+  ApolloProvider,
+  gql,
+  graphql,
+} from 'react-apollo';
 
-export default ({ data: { viewer } }) => {
-  if (viewer) {
-    return <div>{viewer.gists.edges[0].node.id}</div>;
+const networkInterface = createNetworkInterface({
+  uri: 'https://api.github.com/graphql',
+  opts: {
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': `bearer ${process.env.TOKEN}`,
+    },
+  },
+})
+
+const client = new ApolloClient({
+  networkInterface: networkInterface,
+});
+
+const listQuery = gql`
+query {
+  viewer {
+    gists(first: 1) {
+      edges {
+        node {
+          id
+        }
+      }
+    }
   }
-
-  return <div>loading...</div>;
 }
+`;
+
+const ListItem = ({ data: { viewer, loading, error } }) => {
+    if (loading) {
+      return <div>loading...</div>;
+    }
+    if (error) {
+      return <div>error!</div>;
+    }
+    return <div>{viewer.gists.edges[0].node.id}</div>;
+}
+
+const ListItemWithData = graphql(listQuery)(ListItem);
+
+class List extends React.Component {
+  render() {
+    return (
+      <ApolloProvider client={client}>
+        <ListItemWithData />
+      </ApolloProvider>
+    )
+  }
+}
+
+export default List;
